OG-691: Naming convention: unify getter/setter names (#760)
OG-691: Naming convention: unify getter/setter names

Also, make all 'uint' to be 'uint256'.
Rename 'getHubAddress' to 'getRelayHub'.

+++ packages/contracts/src/arbitrum/ArbSys.sol
@@ -9,5 +9,5 @@ interface ArbSys {
      * @notice get the caller's amount of available storage gas
      * @return amount of storage gas available to the caller
      */
-    function getStorageGasAvailable() external view returns (uint);
+    function getStorageGasAvailable() external view returns (uint256);
 }
